channel-api/requirements.txt
apispec==3.3.1
apispec-webframeworks==0.5.2
attrs==20.2.0
base58==1.0.3
bitarray==1.2.2
boto3==1.15.16
botocore==1.18.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
Flask==1.1.2
Flask-Script==2.0.6
hexbytes==0.2.1
idna==2.10
iniconfig==1.1.1
ipfshttpclient==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
libtrustbridge @ git+https://github.com/trustbridge/libtrustbridge@d2befd54b319e4803048bd00e334c5487bab4b05
lru-dict==1.1.6
MarkupSafe==1.1.1
marshmallow==3.6.0
minio==4.0.16
morphys==1.0
multiaddr==0.0.9
multihash==0.1.1
netaddr==0.8.0
openapi-spec-validator==0.2.9
packaging==20.4
parsimonious==0.8.1
pluggy==0.13.1
prance==0.19.0
protobuf==3.13.0
py==1.9.0
py-cid==0.2.1
py-multibase==1.0.1
py-multicodec==0.1.3
py-multihash==0.2.3
pycountry==19.8.18
pycryptodome==3.9.8
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
python-baseconv==1.2.2
python-box==5.1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rlp==2.0.0
rusty-rlp==0.1.15
s3transfer==0.3.3
semver==2.10.2
sentry-sdk==0.14.3
six==1.15.0
toml==0.10.1
toolz==0.11.1
urllib3==1.25.10
varint==1.0.2
web3==5.12.2
websockets==8.1
Werkzeug==1.0.1
